+++ CHANGELOG.md
@@ -12,7 +12,7 @@ March 2023
 
 - Add download menu to chart edit mode view.
 - Add name field to units when exporting to MilX format.
-- Add support for custom symbol fill colors for MilX export.
+- Add support for custom symbol fill colors for MilX import and export.
 - Add custom symbol fill color support for groups.
 - Add custom symbol fill color support for sides.
 - Add setting for selecting simple status modifier.